use float64 for all json numeric properties

+++ ext/lightspuntual/lightspuntual.go
@@ -50,12 +50,12 @@ type LightIndex uint32
 
 // Spot defines the spot cone.
 type Spot struct {
-	InnerConeAngle float32  `json:"innerConeAngle,omitempty"`
-	OuterConeAngle *float32 `json:"outerConeAngle,omitempty"`
+	InnerConeAngle float64  `json:"innerConeAngle,omitempty"`
+	OuterConeAngle *float64 `json:"outerConeAngle,omitempty"`
 }
 
 // OuterConeAngleOrDefault returns the OuterConeAngle if it is not nil, else return the default one.
-func (s *Spot) OuterConeAngleOrDefault() float32 {
+func (s *Spot) OuterConeAngleOrDefault() float64 {
 	if s.OuterConeAngle == nil {
 		return math.Pi / 4
 	}
@@ -81,14 +81,14 @@ type Lights []*Light
 type Light struct {
 	Type      string      `json:"type"`
 	Name      string      `json:"name,omitempty"`
-	Color     *[3]float32 `json:"color,omitempty" validate:"omitempty,dive,gte=0,lte=1"`
-	Intensity *float32    `json:"intensity,omitempty"`
-	Range     *float32    `json:"range,omitempty"`
+	Color     *[3]float64 `json:"color,omitempty" validate:"omitempty,dive,gte=0,lte=1"`
+	Intensity *float64    `json:"intensity,omitempty"`
+	Range     *float64    `json:"range,omitempty"`
 	Spot      *Spot       `json:"spot,omitempty"`
 }
 
 // IntensityOrDefault returns the itensity if it is not nil, else return the default one.
-func (l *Light) IntensityOrDefault() float32 {
+func (l *Light) IntensityOrDefault() float64 {
 	if l.Intensity == nil {
 		return 1
 	}
@@ -96,9 +96,9 @@ func (l *Light) IntensityOrDefault() float32 {
 }
 
 // ColorOrDefault returns the color if it is not nil, else return the default one.
-func (l *Light) ColorOrDefault() [3]float32 {
+func (l *Light) ColorOrDefault() [3]float64 {
 	if l.Color == nil {
-		return [3]float32{1, 1, 1}
+		return [3]float64{1, 1, 1}
 	}
 	return *l.Color
 }
@@ -106,7 +106,7 @@ func (l *Light) ColorOrDefault() [3]float32 {
 // UnmarshalJSON unmarshal the light with the correct default values.
 func (l *Light) UnmarshalJSON(data []byte) error {
 	type alias Light
-	tmp := alias(Light{Color: &[3]float32{1, 1, 1}, Intensity: gltf.Float(1), Range: gltf.Float(float32(math.Inf(0)))})
+	tmp := alias(Light{Color: &[3]float64{1, 1, 1}, Intensity: gltf.Float(1), Range: gltf.Float(math.Inf(0))})
 	err := json.Unmarshal(data, &tmp)
 	if err == nil {
 		*l = Light(tmp)
